+++ tracer/test/Datadog.Trace.Security.IntegrationTests/IAST/IastInstrumentationUnitTests.cs
@@ -162,9 +162,7 @@ public class IastInstrumentationUnitTests : TestHelper
     public void TestConcatMethodsAspectCover()
     {
         var overloadsToExclude = new List<string>() { "System.String Concat(System.Object)" };
-        var typesToExclude = new List<Type>() { typeof(ReadOnlySpan<char>) };
-
-        TestMethodOverloads("System.String", "Concat", overloadsToExclude, typesToExclude);
+        TestMethodOverloads("System.String", "Concat", overloadsToExclude, _notInstrumentedTypes);
     }
 
     [SkippableFact]
@@ -229,7 +227,7 @@ public class IastInstrumentationUnitTests : TestHelper
     private void TestMethodOverloads(string typeToCheck, string methodToCheck, List<string> overloadsToExclude, List<Type> typesToExclude)
     {
         var overloadsToExcludeNormalized = overloadsToExclude?.Select(NormalizeName).ToList();
-        var aspects = Datadog.Trace.ClrProfiler.AspectDefinitions.Aspects.ToList();
+        var aspects = ClrProfiler.AspectDefinitions.Aspects.ToList();
         var type = Type.GetType(typeToCheck);
         type.Should().NotBeNull();
         var typeMethods = type?.GetMethods().Where(x => x.Name == methodToCheck);
