+++ tracer/test/test-applications/integrations/Samples.InstrumentedTests/Vulnerabilities/String/StringSubstringTests.cs
@@ -1,5 +1,4 @@
 using System;
-using System.Security.Cryptography;
 using FluentAssertions;
 using Xunit;
 
@@ -145,7 +144,7 @@ public class StringSubstringTests : InstrumentationTestsBase
     }
 
     [Fact]
-     public void GivenATaintedString_WhenCallingSubstringWithTwoWrongParameter_ThenArgumentOutOfRangeExceptionIsThrown2()
+    public void GivenATaintedString_WhenCallingSubstringWithTwoWrongParameter_ThenArgumentOutOfRangeExceptionIsThrown2()
     {
         Assert.Throws<ArgumentOutOfRangeException>(() => taintedValue.Substring(2200, 1));
     }
