+++ tracer/test/test-applications/integrations/Samples.InstrumentedTests/Vulnerabilities/StringBuilder/StringBuilderAppendTests.cs
@@ -1,7 +1,5 @@
 using System;
-using System.Linq;
 using System.Text;
-using FluentAssertions;
 using Xunit;
 namespace Samples.InstrumentedTests.Iast.Vulnerabilities.StringBuilderPropagation;
 
@@ -24,192 +22,6 @@ public class StringBuilderAppendTests : InstrumentationTestsBase
         AddTainted(TaintedStringBuilder);
         AddTainted(TaintedStringBuilder2);
         AddTainted(TaintedCharArray);
-    }
-
-    // System.Text.StringBuilder::.ctor(System.String)
-
-    [Fact]
-    public void GivenATaintedString_WhenCreateStringBuilder_ResultIsTainted()
-    {
-        AssertTaintedFormatWithOriginalCallCheck(":+-tainted-+:", new StringBuilder(taintedValue), () => new StringBuilder(taintedValue));
-    }
-
-    [Fact]
-    public void GivenATaintedString_WhenCreateStringBuilder_ResultIsTainted2()
-    {
-        AssertTaintedFormatWithOriginalCallCheck("UntaintedString:+-tainted-+:UntaintedString", 
-            new StringBuilder(UntaintedString + taintedValue + UntaintedString), 
-            () => new StringBuilder(UntaintedString + taintedValue + UntaintedString));
-    }
-
-    [Fact]
-    public void GivenATaintedString_WhenCreateStringBuilder_ResultIsNotTainted3()
-    {
-        AssertUntaintedWithOriginalCallCheck("UntaintedString",
-            new StringBuilder(UntaintedString),
-            () => new StringBuilder(UntaintedString));
-    }
-
-    [Fact]
-    public void GivenATaintedString_WhenCreateStringBuilder_ResultIsNotTainted4()
-    {
-        AssertUntaintedWithOriginalCallCheck(string.Empty, new StringBuilder(null), () => new StringBuilder(null));
-    }
-
-    // System.Text.StringBuilder::.ctor(System.String,System.Int32)
-
-    [Fact]
-    public void GivenATaintedString_WhenCreateStringBuilderCapacity_ResultIsTainted()
-    {
-        AssertTaintedFormatWithOriginalCallCheck(":+-tainted-+:", new StringBuilder(taintedValue, 100), () => new StringBuilder(taintedValue, 100));
-    }
-
-    [Fact]
-    public void GivenATaintedString_WhenCreateStringBuilderCapacity_ResultIsTainted2()
-    {
-        AssertTaintedFormatWithOriginalCallCheck("UntaintedString:+-tainted-+:UntaintedString",
-            new StringBuilder(UntaintedString + taintedValue + UntaintedString, 100),
-            () => new StringBuilder(UntaintedString + taintedValue + UntaintedString, 100));
-    }
-
-    [Fact]
-    public void GivenATaintedString_WhenCreateStringBuilderCapacity_ResultIsNotTainted3()
-    {
-        AssertUntaintedWithOriginalCallCheck("UntaintedString",
-            new StringBuilder(UntaintedString, 100),
-            () => new StringBuilder(UntaintedString, 100));
-    }
-
-    [Fact]
-    public void GivenATaintedString_WhenCreateStringBuilderCapacity_ResultIsNotTainted4()
-    {
-        AssertUntaintedWithOriginalCallCheck(string.Empty, new StringBuilder(null, 3), () => new StringBuilder(null, 3));
-    }
-
-    [Fact]
-    public void GivenATaintedString_WhenCreateStringBuilderCapacity_ResultIsNotTainted5()
-    {
-        AssertUntaintedWithOriginalCallCheck("UntaintedString",
-            new StringBuilder(UntaintedString, 1),
-            () => new StringBuilder(UntaintedString, 1));
-    }
-
-    [Fact]
-    public void GivenATaintedString_WhenCreateStringBuilderCapacity_ExceptionIsThrown()
-    {
-        Assert.Throws<ArgumentOutOfRangeException>(() => new StringBuilder(taintedValue, -1));
-    }
-
-    // System.Text.StringBuilder::.ctor(System.String,System.Int32,System.Int32,System.Int32)
-
-    [Fact]
-    public void GivenATaintedString_WhenCreateStringBuilderCapacityOffsets_ResultIsTainted()
-    {
-        AssertTaintedFormatWithOriginalCallCheck(":+-tainted-+:", new StringBuilder(taintedValue, 0, 7, 10), () => new StringBuilder(taintedValue));
-    }
-
-    [Fact]
-    public void GivenATaintedString_WhenCreateStringBuilderCapacityOffsets_ResultIsTainted2()
-    {
-        AssertTaintedFormatWithOriginalCallCheck(":+-in-+:", 
-            new StringBuilder(taintedValue, 2, 2, 10), 
-            () => new StringBuilder(taintedValue, 2, 2, 10));
-    }
-
-    [Fact]
-    public void GivenATaintedString_WhenCreateStringBuilderCapacityOffsets_ResultIsTainted3()
-    {
-        AssertTaintedFormatWithOriginalCallCheck("intedString:+-tain-+:",
-            new StringBuilder(UntaintedString + taintedValue + UntaintedString, 4, 15, 22),
-            () => new StringBuilder(UntaintedString + taintedValue + UntaintedString, 4, 15, 22));
-    }
-
-    [Fact]
-    public void GivenATaintedString_WhenCreateStringBuilderCapacityOffsets_ResultIsNotTainted3()
-    {
-        AssertUntaintedWithOriginalCallCheck("ta",
-            new StringBuilder(UntaintedString, 2, 2, 2),
-            () => new StringBuilder(UntaintedString, 2, 2, 2));
-    }
-
-    [Fact]
-    public void GivenATaintedString_WhenCreateStringBuilderCapacityOffsets_ExceptionIsThrown()
-    {
-        Assert.Throws<ArgumentOutOfRangeException>(() => new StringBuilder(null, 2, 2, 2));
-    }
-
-    [Fact]
-    public void GivenATaintedString_WhenCreateStringBuilderCapacityOffsets_ExceptionIsThrown2()
-    {
-        Assert.Throws<ArgumentOutOfRangeException>(() => new StringBuilder(taintedValue, -1, 3, 3));
-    }
-
-    [Fact]
-    public void GivenATaintedString_WhenCreateStringBuilderCapacityOffsets_ExceptionIsThrown3()
-    {
-        Assert.Throws<ArgumentOutOfRangeException>(() => new StringBuilder(taintedValue, 1, -3, 3));
-    }
-
-    [Fact]
-    public void GivenATaintedString_WhenCreateStringBuilderCapacityOffsets_ExceptionIsThrown4()
-    {
-        Assert.Throws<ArgumentOutOfRangeException>(() => new StringBuilder(taintedValue, 1, 3, -3));
-    }
-
-    [Fact]
-    public void GivenATaintedString_WhenCreateStringBuilderCapacityOffsets_ExceptionIsThrown5()
-    {
-        Assert.Throws<ArgumentOutOfRangeException>(() => new StringBuilder(taintedValue, 1, 300, 3));
-    }
-
-    [Fact]
-    public void GivenATaintedString_WhenCreateStringBuilderCapacityOffsets_ExceptionIsThrown6()
-    {
-        Assert.Throws<ArgumentOutOfRangeException>(() => new StringBuilder(taintedValue, 100, 3, 3));
-    }
-
-    // System.Object::ToString()
-
-    [Fact]
-    public void GivenATaintedString_WhenCallingStringEmptyBuilderAppendAppendAppendToString_ResultIsTainted()
-    {
-        AssertTaintedFormatWithOriginalCallCheck("pre:+-tainted-+:post", new StringBuilder().Append("pre").Append(taintedValue).Append("post").ToString().ToString(), () => new StringBuilder().Append("pre").Append(taintedValue).Append("post").ToString().ToString());
-    }
-
-    // System.Text.StringBuilder::ToString(System.Int32,System.Int32)
-
-    [Fact]
-    public void GivenATaintedString_WhenCallingStringBuilderToString_ResultIsTainted()
-    {
-        AssertTaintedFormatWithOriginalCallCheck(":+-ta-+:",
-            new StringBuilder(taintedValue).ToString(0, 2),
-            () => new StringBuilder(taintedValue).ToString(0, 2));
-    }
-
-    [Fact]
-    public void GivenATaintedString_WhenCallingStringBuilderToString_ResultIsTainted2()
-    {
-        AssertTaintedFormatWithOriginalCallCheck(":+-in-+:",
-            new StringBuilder(taintedValue).ToString(2, 2),
-            () => new StringBuilder(taintedValue).ToString(2, 2));
-    }
-
-    [Fact]
-    public void GivenATaintedString_WhenCallingStringBuilderToString_ResultIsTainted3()
-    {
-        Assert.Throws<ArgumentOutOfRangeException>(() => new StringBuilder(taintedValue).ToString(-1, 2));
-    }
-
-    [Fact]
-    public void GivenATaintedString_WhenCallingStringBuilderToString_ResultIsTainted4()
-    {
-        Assert.Throws<ArgumentOutOfRangeException>(() => new StringBuilder(taintedValue).ToString(1, -2));
-    }
-
-    [Fact]
-    public void GivenATaintedString_WhenCallingStringBuilderToString_ResultIsTainted5()
-    {
-        Assert.Throws<ArgumentOutOfRangeException>(() => new StringBuilder(taintedValue).ToString(1, 332));
     }
 
     // System.Text.StringBuilder::Append(System.String)
@@ -264,6 +76,12 @@ public class StringBuilderAppendTests : InstrumentationTestsBase
         AssertTaintedFormatWithOriginalCallCheck(":+-tainted-+:",
             new StringBuilder(taintedValue).Append((string)null).ToString(),
             () => new StringBuilder(taintedValue).Append((string)null).ToString());
+    }
+
+    [Fact]
+    public void GivenATaintedString_WhenCallingStringEmptyBuilderAppendAppendAppendToString_ResultIsTainted()
+    {
+        AssertTaintedFormatWithOriginalCallCheck("pre:+-tainted-+:post", new StringBuilder().Append("pre").Append(taintedValue).Append("post").ToString().ToString(), () => new StringBuilder().Append("pre").Append(taintedValue).Append("post").ToString().ToString());
     }
 
     [Fact]
